requirements.txt
dj-database-url==0.5.0
Django==2.2
django-cors-headers==2.5.2
django-extensions==2.1.6
djangorestframework==3.9.2
factory-boy==2.11.1
Faker==1.0.4
gunicorn==19.9.0
psycopg2==2.7.7
pygraphviz==1.5
python-dateutil==2.8.0
pytz==2018.9
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
whitenoise==4.1.2
